+++ src/model/modelCNN.py
